+++ src/otx/backend/native/models/detection/backbones/hgnetv2.py
@@ -626,6 +626,43 @@ class HGNetv2:
             "freeze_norm": True,
             "freeze_stem_only": True,
         },
+        "deim_dfine_hgnetv2_n": {
+            "name": "B0",
+            "return_idx": [2, 3],
+            "freeze_at": -1,
+            "freeze_norm": False,
+            "use_lab": True,
+            "freeze_stem_only": True,
+            "pretrained": True,
+        },
+        "deim_dfine_hgnetv2_s": {
+            "name": "B0",
+            "return_idx": [1, 2, 3],
+            "freeze_at": -1,
+            "freeze_norm": False,
+            "use_lab": True,
+        },
+        "deim_dfine_hgnetv2_m": {
+            "name": "B2",
+            "return_idx": [1, 2, 3],
+            "freeze_at": -1,
+            "freeze_norm": False,
+            "use_lab": True,
+        },
+        "deim_dfine_hgnetv2_l": {
+            "name": "B4",
+            "return_idx": [1, 2, 3],
+            "freeze_at": 0,
+            "freeze_norm": True,
+            "freeze_stem_only": True,
+        },
+        "deim_dfine_hgnetv2_x": {
+            "name": "B5",
+            "return_idx": [1, 2, 3],
+            "freeze_at": -1,
+            "freeze_norm": False,
+            "freeze_stem_only": True,
+        },
     }
 
     def __new__(cls, model_name: str) -> HGNetv2Module: